never exit w/a 0 status code if any excepts caught
I was trying to iterate over files and use the exit code from this script to find out if my json files were valid, however I realized the script always returns 0. Even if the json is unparseable, `raise SystemExit` on L45 implicitly exits w/0 and in the case of an actual failed validation, no exceptions are raised and so the script exits w/a 0. This change, I think, let's the user distinguish between a good validation (0), an unparseable json file (1), and a failed validation (2). I was going to try and make the status code equal to the failed line number in the latter case, but that's beyond my skill and patience at the moment.