Add files via upload
This is a file that was created at the 5/6/2025 AWG meeting to contain all of the validation rules that CVE Services enforces as part of the CVE Record Publication process.    The idea is that this file will be updated every time that a new "rule" is enforced by CVE Services when CNA publish CVE records.